requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
python-dotenv==0.17.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.15.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
